refactor(cli): depend only on server
