+++ crates/mm-cli/Cargo.toml
@@ -5,7 +5,6 @@ edition = "2024"
 default-run = "mm-cli"
 
 [dependencies]
-mm-core = { path = "../mm-core" }
 mm-server = { path = "../mm-server" }
 clap = { version = "4.4", features = ["derive"] }
 tokio = { workspace = true, features = ["full"] }
